Use 64-bit-safe format specifiers for logged file offsets and control flags
On Windows (LLP64) `long` is 32-bit, so `%ld`/`%lx` truncate or mis-read the
64-bit values these diagnostics pass:

- dump_file.h:243 logged a 64-bit ftell() file offset via a `(long)` cast, which
  silently truncates offsets >= 2 GiB on Windows.
- dump_file.h:256 passed a 64-bit offset straight to `%ld` (a varargs size
  mismatch -> undefined behavior on Windows).
- meta_service.cpp:561 logged two int64_t control-flag fields with `%lx`; on
  Windows each consumes only 32 bits and misaligns the remaining varargs.

Print them with PRId64/PRIx64 and fixed-width int64_t/uint64_t casts (matching
server_state.cpp in the same directory) and include <cinttypes>. Diagnostic
output only; behavior on LP64 (Linux/macOS) is unchanged.

+++ src/meta_server/meta_server_lib/dump_file.h
@@ -41,6 +41,7 @@
 #include <cerrno>
 #include <iostream>
 #include <exception>
+#include <cinttypes>
 
 inline void error_msg(int err_number, /*out*/char* buffer, int buflen)
 {
@@ -240,7 +241,8 @@ public:
             {
                 if ( feof(_file_handle) )
                 {
-                    derror("unexpected file end, start offset of this block (%ld)", (long)(ftell(_file_handle)-len-sizeof(hdr)));
+                    derror("unexpected file end, start offset of this block (%" PRId64 ")",
+                           static_cast<int64_t>(ftell(_file_handle)) - static_cast<int64_t>(len) - static_cast<int64_t>(sizeof(hdr)));
                     return -1;
                 }
                 else if (errno != EINTR)
@@ -253,7 +255,8 @@ public:
         _crc = dsn_crc32_compute(raw_mem, len, _crc);
         if (_crc != hdr.crc32)
         {
-            derror("file %s data error, block offset(%ld)", _filename.c_str(), ftell(_file_handle)-hdr.length-sizeof(hdr));
+            derror("file %s data error, block offset(%" PRId64 ")", _filename.c_str(),
+                   static_cast<int64_t>(ftell(_file_handle)) - static_cast<int64_t>(hdr.length) - static_cast<int64_t>(sizeof(hdr)));
             return -1;
         }
